+++ Assets/Scripts/Battlefield/AbstractTurnBrain.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9c7abfd91b821ad488269438a56f56f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
