--- Assets/Scripts/Battlefield/CreatureScripts/AAbilitySO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 013f77162c20cf240b3b4f7baba1c77d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
