+++ Assets/Scripts/Battlefield/CreatureScripts/UnitAbilitiesContainer.cs
@@ -7,13 +7,13 @@ namespace SwordAndBored.Battlefield.CreaturScripts
     public class UnitAbilitiesContainer : MonoBehaviour
     {
         [Header("Ability Info")]
-        public AAbilitySO[] abilities;
+        public AbstractAbilitySO[] abilities;
         UniqueCreature unit;
 
         void Start()
         {
             unit = GetComponent<UniqueCreature>();
-            foreach (AAbilitySO ability in abilities)
+            foreach (AbstractAbilitySO ability in abilities)
             {
                 ability.Initialize(transform.gameObject);
             }
@@ -22,30 +22,8 @@ namespace SwordAndBored.Battlefield.CreaturScripts
         public void UseAbility(int i, GameObject target)
         {
             unit.animator.SetTrigger("Attack");
-            switch (abilities[i].TypeOfActionRequired)
-            {
-                case AAbilitySO.ActionTypes.Action:
-                    if (unit.action)
-                    {
-                        abilities[i].TriggerAbility(target);
-                        unit.action = false;
-                    }
-                    break;
-                case AAbilitySO.ActionTypes.BonusAction:
-                    if (unit.bonus)
-                    {
-                        abilities[i].TriggerAbility(target);
-                        unit.bonus = false;
-                    }
-                    break;
-                case AAbilitySO.ActionTypes.Reaction:
-                    if (unit.reaction)
-                    {
-                        abilities[i].TriggerAbility(target);
-                        unit.reaction = false;
-                    }
-                    break;
-            }
+            abilities[i].TriggerAbility(target);
+            unit.action = false;
         }
     }
