+++ Assets/Scripts/Battlefield/TurnBrain.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace SwordAndBored.Battlefield
+{
+    public class TurnBrain : MonoBehaviour
+    {
+        public AbstractTurnBrain brain;
+    
+        void Update()
+        {
+            brain.UpdateBrainHandler();
+        }
+    }
+
+}
